Adiciona uma estrutura de controle de fluxo, try-catch | adiociona switch casa para lidar com interação do usuario
